Add a callback to be invoked before retrying a request
When using an `io.Reader` as the request body, it's possible for some of the
body to be read by the initial, unauthenticated attempt. When this happens, the
authenticated retry attempt will not transmit the full body.

To accommodate this, add a callback to the request and invoke it before
performing a retry. This gives callers the opportunity to make any necessary
adjustments to the body (or other parts of the request) before the request is
retried.

Signed-off-by: Adam Wolfe Gordon <awg@digitalocean.com>